continuous refresh for MemoryMaps
