+++ collect/etw/MicrosoftWindowsCpuTriggerCollector.cs
@@ -81,60 +81,5 @@ namespace gov.llnl.wintap.collect
             }
          
         }
-
-        //public override void Process_Event(TraceEvent obj)
-        //{
-        //    // 1.)  Call base to setup rate monitoring
-        //    base.Process_Event(obj);
-        //    try
-        //    {
-        //        if(obj.EventName == "CpuUsageEvent")
-        //        {
-        //            // 2.)  Process the event
-        //            Dictionary<string, string> parsedSessionEvent = parseEvent(obj.ToString());
-        //            CpuUsageEventData metric = new CpuUsageEventData();
-        //            this.CollectorName = this.GetType().Name; // resetting to the classname so that dynamic consumers can instantiate the model object
-        //            metric.OnBatteryPower = StateManager.OnBatteryPower;
-        //            metric.UserBusy = StateManager.UserBusy;
-        //            metric.UserName = StateManager.ActiveUser;
-        //            metric.ComputerName = Environment.MachineName;
-        //            int pid = Convert.ToInt32(parsedSessionEvent["ProcessId"]);
-        //            try
-        //            {
-        //                metric.AppPath = this.TranslateFilePath(parsedSessionEvent["ImagePath"]);
-        //            }
-        //            catch(Exception ex)
-        //            {
-        //                WintapLogger.Log.Append("CpuTriggerCollector Could not translate ImagePath. Path: " + parsedSessionEvent["ImagePath"] + "  error: " + ex.Message, LogLevel.Always);
-        //            }
-
-        //            metric.AppName = parsedSessionEvent["ProcessInfo"];
-        //            metric.AppDescription = metric.AppName;
-        //            metric.PID = pid;
-        //            metric.Timestamp = obj.TimeStamp;
-        //            double cpuTimeMs = Convert.ToInt64(parsedSessionEvent["CpuTimeIncrementInMs"]);
-        //            int intervalMs = Convert.ToInt32(parsedSessionEvent["TimeSinceLastCheckInMs"]);
-        //            int cpuCount = Environment.ProcessorCount;
-        //            double totalProcessorTimeMs = cpuCount * intervalMs;
-        //            double percentCpu = cpuTimeMs / totalProcessorTimeMs;
-
-        //            double rawPercent = percentCpu * 100;
-        //            metric.AppCpuPercentage = (int)rawPercent;
-
-        //            // 3.) Create a WintapMessage and attach your event to it
-        //            WintapMessage wintapMsg = new WintapMessage(obj.TimeStamp, metric.PID, "CpuUsageEvent");
-        //            wintapMsg.ActivityType = obj.EventName;
-        //            wintapMsg.CpuUsageEvent = metric;
-
-        //            // 4.) Send your event into the Wintap event pipeline
-        //            wintapMsg.Send();
-        //        }
-        //    }
-        //    catch (Exception ex)
-        //    {
-        //        WintapLogger.Log.Append("Error processing " + this.CollectorName + " event: " + ex.Message, LogLevel.Debug);
-
-        //    }
-        //}
     }
 }
